advanced_rag/requirements.txt
langchain == 0.3.0
openai == 1.55.3
azure-search-documents == 11.6.0b2
azure-identity == 1.19.0
pypdf == 4.3.1
streamlit == 1.37.1
python-dotenv == 1.0.1
